API: Changes in render API
Instead of output, the parameter is now named request
and it's the last parameter in hoedown_document_render.

The request parameter is now mainly a per-render input
parameter, that can be accessed at any time via data->request.

render_end now directly returns the output, which in turn
is returned by hoedown_document_render to the user.
This makes for a simpler API.

All global callbacks are required now.