Infer is_tight from context when rendering paragraphs
Previously, the document parser had to propagate is_tight
across paragraphs so we could render paragraphs inside
tight lists correctly.

Now that we are passing context to the renderers, the hack is
no longer needed; the renderer can "mark" objects of tight list
items and render their paragraphs differently based on that.